requirements.txt
fastapi
passlib[bcrypt]
uvicorn
pymongo
python-multipart
python-jose[cryptography]
pydantic[email]
python-dotenv

